+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+-----------------------
+##### Fixed
+- An exception occurs when a repeat is deleted.
+
 [5.16.12] - 2020-12-04
 -----------------------
 ##### Changed